Automatically set proper popup position and size
If 'wrap' is enabled, Vim won't shift the popup window when it doesn't
fit onto the screen. If 'wrap' is disabled, Vim will properly shift the
popup to the left but very long lines will be cut off. Therefore, we set
'col', 'minwidth' and 'maxwidth' manually, and keep 'wrap' enabled.

With these settings the popup is shifted as much to the left as
necessary to show all text lines in the popup unwrapped. When the text
lines are longer than the screen width, the popup window will be set to
the width of the window width and the text in the popup will be wrapped.